backend/pyproject.toml
[tool.poetry]
name = "digit"
version = "0.1.0"
description = "Print smart sensors monthly reports using data from the ABB cloud"
authors = ["Davide Silvestri <silvestri.eng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.0.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.0.0"
requests = "^2.27.1"
djangorestframework-dataclasses = "^1.1.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "21.7b0"
pytest = "^6.2.5"
flake8 = "^4.0.1"
pytest-django = "4.4.0"
pytest-env = "0.6.2"
pytest-asyncio = "0.15.1"
pytest-ordering = "0.6"
pytest-mock = "3.6.1"
freezegun = "1.1.0"
responses = "0.13.4"
watchdog = "2.1.4"
argh = "0.26.2"
pyinstrument = "4.0.3"
pyfakefs = "4.5.0"
pytest-xdist = "2.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
